move the clock thing to it's own file
